# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: JobPing API
  description: |
    JobPing is a sophisticated job matching platform for early-career professionals.
    This API provides comprehensive job matching, user management, and analytics capabilities.
    
    ## Key Features
    - AI-powered job matching with GPT-4
    - Rule-based fallback matching
    - Real-time job scraping from multiple sources
    - User preference management
    - Email automation and notifications
    - Payment processing with Stripe
    - Comprehensive analytics and monitoring
    
    ## Authentication
    Most endpoints require authentication via JWT tokens or API keys.
    Public endpoints are clearly marked.
    
    ## Rate Limiting
    API requests are rate-limited per endpoint to ensure fair usage:
    - Public endpoints: 20 requests/minute
    - Authenticated endpoints: 100 requests/minute
    - Resource-intensive endpoints: 2-5 requests/minute
    
    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
    See the Error Response schema for details.
  version: 1.0.0
  contact:
    name: JobPing Support
    email: support@jobping.ai
    url: https://jobping.ai/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://jobping.ai/api
    description: Production server
  - url: https://staging.jobping.ai/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
...
paths:
...
  /user-matches:
    get:
      tags:
        - Job Matching
      summary: Get User Matches
      description: |
        Retrieves job matches for a specific user with pagination and filtering options.
      operationId: getUserMatches
      security:
        - BearerAuth: []
      parameters:
        - name: email
          in: query
          required: true
          description: User email address
          schema:
            type: string
            format: email
        - name: page
          in: query
          required: false
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Number of matches per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: minScore
          in: query
          required: false
          description: Minimum match score filter
          schema:
            type: integer
            minimum: 0
            maximum: 100
            default: 50
      responses:
        '200':
          description: User matches retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  matches:
                    type: array
                    items:
                      $ref: '#/components/schemas/JobMatch'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  totalMatches:
                    type: integer
                    example: 45
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for system authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
  schemas:
...
    JobMatch:
      type: object
      properties:
        job_index:
          type: integer
          description: Index of the job in the original array
          example: 0
        job_hash:
          type: string
          description: Unique identifier for the job
          example: "abc123def456"
        match_score:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall match score
          example: 85
        match_reason:
          type: string
          description: Human-readable explanation of the match
          example: "Strong skill alignment, great location match"
        confidence_score:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence in the match quality
          example: 0.87
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 5
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Invalid email format"
        code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: object
          additionalProperties: true
        timestamp:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"
        requestId:
          type: string
          example: "req_1234567890"
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            message: "Authentication required"
            code: "UNAUTHORIZED"
            timestamp: "2024-01-15T10:30:00Z"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Internal Server Error"
            message: "An unexpected error occurred"
            code: "INTERNAL_ERROR"
            timestamp: "2024-01-15T10:30:00Z"